fix: resolve all ruff linting errors and frontend test failures (#167)
* fix: resolve all ruff linting errors and frontend test failures

Backend:
- Fix undefined names in financials.py (AccountUpdate, DUTCH_RGS_LIGHT)
- Add _project_for_user_or_404 helper for cost endpoints
- Add missing import for customer email notifications
- Consolidate mid-file imports to top of financials.py
- Fix ambiguous Unicode chars (en-dash, multiplication sign)
- Fix line length violations
- Update ruff config: relax TCH rules for runtime-dependent modules,
  ignore false-positive style rules that conflict with FastAPI patterns
- Auto-fix import sorting, unused imports, enum modernization

Frontend:
- Fix push-permission.test.tsx: rewrite to match actual component API
- Fix push-notifications.test.tsx: keep PushManager alive through async renders
- Fix auth-refresh.test.tsx: add missing component mocks (offline-indicator,
  pwa-register, mobile-nav, mobile-time-tracker)
- Fix dashboard.test.tsx: mock apiFetch for Settings page to prevent
  unhandled ECONNREFUSED rejections

Results: ruff 0 errors, vitest 781/781 pass, pytest 729 pass (13 pre-existing)


* fix: update CI workflow push condition


* chore: trigger CI re-run

* fix: replace reusable workflows with inline CI jobs

Reusable workflows from gitops-templates cause startup_failure.
Inline all CI jobs to unblock the pipeline.


* fix: resolve ruff formatting and TypeScript errors for CI

- Auto-format 77 backend Python files with ruff
- Fix frontend TypeScript type mismatches in types.ts, processes page,
  time-tracking page, agenda, push, and reviews modules
- Fix tsconfig strictness settings


* fix: update time-tracking-widget tests to match refactored TimeTracker API

TimeTracker now accepts projectId and fetches processes internally.
Update test mock and assertions to match current component interface.


* fix: add beautifulsoup4 to backend dependencies

The bouwmaat store scraper imports bs4 (BeautifulSoup) but the package
was missing from pyproject.toml, causing ImportError in CI.


* fix: pin Python 3.12 in CI for consistent uv resolution


* fix: commit invoice-helpers.ts missing from tracked files

The formatMoney and formatInvoiceDate helpers were imported by the
invoice detail page but never committed, causing Next.js build failure.


* fix: remove non-page exports from invoice page, fix stale security test

- Move formatMoney/formatInvoiceDate to lib/invoice-helpers (already existed)
  and import from there — Next.js rejects non-page exports from page files
- Update test_security.py to expect ValueError on type mismatch (matches
  current decode_token implementation with type validation)


* fix: wrap useSearchParams in Suspense boundary for agenda/day page

Next.js 15 requires useSearchParams() to be inside a Suspense boundary
during static prerendering. Split into wrapper + content component.


* fix: fix stale refresh token test, split agenda/day for Suspense SSG

- test_security.py: pass expected_type="refresh" when decoding refresh tokens
  (matches decode_token's type validation implementation)
- Split agenda/day page into server component (page.tsx with Suspense) and
  client component (agenda-day-content.tsx with useSearchParams) to fix
  Next.js static generation error


* fix: set Docker build context to src/backend and src/frontend

The Dockerfiles use paths relative to their own directories.
Context must match the Dockerfile's directory.


* fix: Docker builds use docker driver, fix refresh token test

- Switch Docker buildx from docker-container to docker driver to fix
  context transfer issues with GHA cache
- Remove stale GHA build cache (was from old repo layout)
- Fix test_decodes_valid_refresh_token to pass expected_type="refresh"


---------

+++ src/backend/app/services/planning/scheduler.py
@@ -36,7 +36,7 @@ class ConstantWeatherProvider:
     def __init__(self, *, bad: bool = False) -> None:
         self._bad = bad
 
-    async def is_bad(self, day_index: int) -> bool:  # noqa: ARG002
+    async def is_bad(self, day_index: int) -> bool:
         return self._bad
 
 
@@ -81,9 +81,7 @@ def _topo_sort_critical_first(tasks: list[CpmTask]) -> list[CpmTask]:
     if not tasks:
         return []
     snapshot = [
-        CpmTask(id=t.id, name=t.name, duration_hours=t.duration_hours,
-                dependencies=list(t.dependencies))
-        for t in tasks
+        CpmTask(id=t.id, name=t.name, duration_hours=t.duration_hours, dependencies=list(t.dependencies)) for t in tasks
     ]
     compute_critical_path(snapshot)
     task_map = {t.id: t for t in snapshot}
@@ -176,7 +174,7 @@ class ScheduleOptimizer:
         scheduled: list[ScheduledTask] = []
         for cpm in ordered:
             si = input_by_id[cpm.id]
-            duration_s = int(round(cpm.duration_hours * _HOUR_S))
+            duration_s = round(cpm.duration_hours * _HOUR_S)
 
             # Dependency lower bound
             dep_finishes = [finish_by_id[d] for d in cpm.dependencies if d in finish_by_id]
@@ -227,12 +225,14 @@ class ScheduleOptimizer:
                 usage.setdefault(res, []).append((start_s, end_s))
 
             finish_by_id[cpm.id] = end_s
-            scheduled.append(ScheduledTask(
-                task_id=cpm.id,
-                start_time_s=start_s,
-                end_time_s=end_s,
-                reasoning="; ".join(reasons),
-            ))
+            scheduled.append(
+                ScheduledTask(
+                    task_id=cpm.id,
+                    start_time_s=start_s,
+                    end_time_s=end_s,
+                    reasoning="; ".join(reasons),
+                )
+            )
 
         # Preserve input task order in output
         order_index = {si.task.id: i for i, si in enumerate(inputs)}
